+++ core/src/main/java/eu/tsystems/mms/tic/testframework/webdrivermanager/WebDriverRequest.java
@@ -1,6 +1,7 @@
 package eu.tsystems.mms.tic.testframework.webdrivermanager;
 
 import java.io.Serializable;
+import java.util.Map;
 
 public interface WebDriverRequest extends Serializable {
     String DEFAULT_SESSION_KEY = "default";
@@ -10,4 +11,5 @@ public interface WebDriverRequest extends Serializable {
     boolean getShutdownAfterTest();
     boolean getShutdownAfterTestFailed();
     boolean getShutdownAfterExecution();
+    Map<String, Object> getCapabilities();
 }
